Implement individual threads for each StatusItem
This commit implements individual threads for each StatusItem defined by
the user in the config file. These threads will run simultaneously and
enqueue the StatusItem text to a concurrent queue when the text has
changed. This prevents redundant information from being written to
stdout as long as no new information is available in any of the
StatusItems.

This architecture also allows for resource-heavy status item scripts
where it may takes a significant amount of time to generate the status
item text, since the resource-heavy operation will not be blocking the
other status item's text generation.